+++ CHANGELOG.md
@@ -12,6 +12,7 @@ CHANGELOG for FlatCAM beta
 - in CNCJob UI added GUI for an eventual Autolevelling feature 
 - in CNCJob UI updated the GUI for Autolevelling
 - Cutout Tool - finished handler for gaps thickness control for the manual gaps
+- CNCJob object - working in generating Voronoi diagram for autolevelling
 
 11.08.2020